requirements.txt
torch>=2.1.1
torchvision>=0.16.1
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.65.0
pyyaml>=6.0
albumentations>=1.3.0
deepface>=0.0.79
